Reorganised the utils
