+++ resolvers/Upcoming/ShowResolver.js
@@ -2,10 +2,8 @@ const axios = require("axios");
 const { has, forEach } = require("lodash");
 
 const { generateUpcomingEndpoint } = require("../../utils/generateEndpoints");
-
 const generateImageURL = require("../../utils/generateImageURL");
-
-const { formatDate } = require("../../utils/formatDates");
+const formatDate = require("../../utils/dates/custom");
 
 const NowPlayingTVResolver = async (parent, args, context, info) => {
   try {
