+++ utils/dates/custom/index.js
@@ -1,0 +1,10 @@
+const formatDate = (date, pattern) => {
+  let formattedDate = moment(date).format(pattern);
+
+  if (isEmpty(date) === true) {
+    formattedDate = "--/--/--";
+    return formattedDate;
+  }
+
+  return formattedDate;
+};
