+++ utils/dates/generateYear/index.js
@@ -1,0 +1,14 @@
+const { isEmpty } = require("lodash");
+
+const generateYear = (date) => {
+  let newYear = moment(date).format("YYYY");
+
+  if (isEmpty(date) === true) {
+    newYear = "-";
+    return newYear;
+  }
+
+  return newYear;
+};
+
+module.exports = generateYear;
